python/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
decorator==4.4.2
distro==1.9.0
elevenlabs==1.11.0
ffmpeg-python==0.2.0
Flask==3.0.3
Flask-Cors==5.0.0
future==1.0.0
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-speech==2.28.0
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
MarkupSafe==3.0.2
moviepy==1.0.3
numpy==2.1.2
openai==1.52.2
pillow==11.0.0
proglog==0.1.10
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
setuptools==75.2.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
Werkzeug==3.0.6
gunicorn